`slug` includes the whole hierarchy and add `name`
The `slug` attribute now includes the whole hierarchy, so that two
sub-pages in different dirs can have the same `name`. Also translation
were incorrectly set, because the `slug` is used by Pelican more or
less like an ID.

So to compute the filename, this commit introduces the `name` attribute
and the `slug` is only used to make Pelican well behaved.

- support all dirs in `PAGE_PATHS`, not only the first
- `_scan_dir_r` supports `base_path` and `rel_path`
- use `_update_context`
- logging tree structure with `title` instead of `slug`s
- fixed a bug that when a temporary `HiPage` is replaced by a real one,
the `parent` attribute of its children wasn't updated correctly.
- HiPages with status "draft" are appended to the `pages`.